Se corrige la relación many-to-many entre usuarios y roles en las entidades UserEntity y RolEntity.

+++ src/main/java/com/exe/ConjuntoResidencialArkania/Impl/RolServiceImpl.java
@@ -470,7 +470,7 @@ public class RolServiceImpl implements RolService {
         return roles.stream()
                 .collect(Collectors.toMap(
                     RolEntity::getNombre,
-                    rol -> rol.getUsuarios().size()
+                    rol -> (Integer) rol.getUsuarios().size()
                 ));
     }
